requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
plotly>=5.18.0
joblib>=1.3.0
fpdf2>=2.7.0
faker>=23.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
nltk>=3.8.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
